+++ src/main.rs
@@ -2,8 +2,12 @@ use std::{thread::sleep, time::Duration};
 
 use anyhow::Result;
 use wlink::{
-    commands, dmi::DebugModuleInterface, firmware::read_firmware_from_file,
-    operations::ProbeSession, probe::WchLink, regs, RiscvChip,
+    commands,
+    dmi::DebugModuleInterface,
+    firmware::{read_firmware_from_file, Firmware},
+    operations::ProbeSession,
+    probe::WchLink,
+    regs, RiscvChip,
 };
 
 use clap::{Parser, Subcommand};
@@ -323,21 +327,38 @@ fn main() -> Result<()> {
                 } => {
                     sess.dump_info()?;
 
-                    let firmware = read_firmware_from_file(path)?;
-                    let start_address =
-                        address.unwrap_or_else(|| sess.chip_family.code_flash_start());
-                    log::info!(
-                        "Flashing {} bytes to 0x{:08x}",
-                        firmware.len(),
-                        start_address
-                    );
-
                     if erase {
                         log::info!("Erase Flash");
                         sess.erase_flash()?;
                     }
 
-                    sess.write_flash(&firmware, start_address)?;
+                    let firmware = read_firmware_from_file(path)?;
+
+                    match firmware {
+                        Firmware::Binary(data) => {
+                            let start_address =
+                                address.unwrap_or_else(|| sess.chip_family.code_flash_start());
+                            log::info!("Flashing {} bytes to 0x{:08x}", data.len(), start_address);
+                            sess.write_flash(&data, start_address)?;
+                        }
+                        Firmware::Sections(sections) => {
+                            // Flash section by section
+                            if address != None {
+                                log::warn!("--address is ignored when flashing ELF or ihex");
+                            }
+                            for section in sections {
+                                let start_address =
+                                    sess.chip_family.fix_code_flash_start(section.address);
+                                log::info!(
+                                    "Flashing {} bytes to 0x{:08x}",
+                                    section.data.len(),
+                                    start_address
+                                );
+                                sess.write_flash(&section.data, start_address)?;
+                            }
+                        }
+                    }
+
                     log::info!("Flash done");
 
                     sleep(Duration::from_millis(500));