+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `--watch-serial` for `flash` subcommand, #36
 - Add `-o/--out` for `dump` sumcommand, #38
+- BREAKING CHANGE: Refactor
+- Add Windows native driver support, #39
 
 ### Changed